backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.4.0
pypdf>=4.3.0
pillow>=10.4.0
pytesseract>=0.3.13
sentence-transformers>=3.0.1
numpy>=1.26.0

